cloud/src/main/java/edu/montana/csci/csci366/strweb/ops/Sha256Transformer.java
package edu.montana.csci.csci366.strweb.ops;

import java.util.concurrent.CountDownLatch;
import java.util.concurrent.Executors;
import java.util.concurrent.ThreadPoolExecutor;

/**
 * This class is should calculate the SHA 256 hexadecimal hash for each line and replace the line in the
 * array with that hash.
 *
 * It should do so using the ThreadPoolExecutor created below.
 */
public class Sha256Transformer {
    String[] _lines;
    ThreadPoolExecutor executor = (ThreadPoolExecutor) Executors.newFixedThreadPool(10);

    public Sha256Transformer(String strings) {
        _lines = strings.split("\n");
    }

    public String toSha256Hashes() {
        //TODO - this method should create a series of Runnables and use the executor to do all
        //       word counts in parallel
      return "";
    }

}
